browser/requirements.txt
blinker==1.9.0
Brotli==1.1.0
click==8.1.8
Flask==3.1.0
Flask-Compress==1.17
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==3.1.3
zstandard==0.23.0
